Warnings, Doxygen

+++ units/DataHolder/main.cpp
@@ -97,7 +97,7 @@ void testClone() {
     holder1->append("B", {2,3,4});
     holder1->append("C", {15});
     holder1->build();
-    for(int i=0; i<holder1->size(); i++)
+    for(size_t i=0; i<holder1->size(); i++)
 	holder1->raw(i)=static_cast<T>(i);
 
     auto holder2=std::make_shared<DataHolder<T>>();
